+++ Package.swift
@@ -9,7 +9,10 @@ let package = Package(
     products: [
         .library(
             name: "KlaviyoSwift",
-            targets: ["KlaviyoSwift"])
+            targets: ["KlaviyoSwift"]),
+        .library(
+            name: "KlaviyoSwiftExtension",
+            targets: ["KlaviyoSwiftExtension"])
     ],
     dependencies: [
         .package(url: "https://github.com/pointfreeco/swift-snapshot-testing", from: "1.10.0"),
@@ -23,7 +26,14 @@ let package = Package(
     targets: [
         .target(
             name: "KlaviyoSwift",
-            dependencies: [.product(name: "AnyCodable", package: "AnyCodable")]),
+            dependencies: [.product(name: "AnyCodable", package: "AnyCodable")],
+            path: "Sources/KlaviyoSwift"
+        ),
+        .target(
+            name: "KlaviyoSwiftExtension",
+            dependencies: [],
+            path: "Sources/KlaviyoSwiftExtension"
+        ),
         .testTarget(
             name: "KlaviyoSwiftTests",
             dependencies: [
